Refact profile edit page on multiple components
